feature(role): Check Role (admin, user)

+++ routes/api.php
@@ -31,20 +31,20 @@ Route::prefix('products')->group(function () {
     Route::get('{slug}', [ProductController::class, 'show']);
     // Product Variants by product_id
     Route::get('{slug}/variants', [ProductVariantController::class, 'byProduct']);
-    Route::middleware('auth:sanctum')->group(function () {
+    Route::middleware(['auth:sanctum', 'role:admin'])->group(function () {
         Route::post('/', [ProductController::class, 'store']);
         Route::put('{slug}', [ProductController::class, 'update']);
         Route::delete('{slug}', [ProductController::class, 'destroy']);
+        Route::post('{slug}/images', [ProductImageController::class, 'addImages']);
+        Route::delete('{slug}/images/{id}', [ProductImageController::class, 'destroy']);
     });
-    Route::post('{slug}/images', [ProductImageController::class, 'addImages']);
-    Route::delete('{slug}/images/{id}', [ProductImageController::class, 'destroy']);
 });
 
 // Product Variant Routes
 Route::prefix('product-variants')->group(function () {
     Route::get('/', [ProductVariantController::class, 'index']);
     Route::get('{id}', [ProductVariantController::class, 'show']);
-    Route::middleware('auth:sanctum')->group(function () {
+    Route::middleware(['auth:sanctum', 'role:admin'])->group(function () {
         Route::post('/', [ProductVariantController::class, 'store']);
         Route::put('{id}', [ProductVariantController::class, 'update']);
         Route::delete('{id}', [ProductVariantController::class, 'destroy']);
@@ -53,4 +53,4 @@ Route::prefix('product-variants')->group(function () {
 });
 
 // Variant Images delete
-Route::delete('product-variant-images/{id}', [ProductVariantImageController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('product-variant-images/{id}', [ProductVariantImageController::class, 'destroy'])->middleware(['auth:sanctum', 'role:admin']);
